Add private extension install method
Had to set private ext files to read only because Android 14 targetting

https: //developer.android.com/about/versions/14/behavior-changes-14#safer-dynamic-code-loading

+++ app/src/main/java/eu/kanade/tachiyomi/ui/extension/ExtensionBottomSheet.kt
@@ -21,6 +21,8 @@ import eu.kanade.tachiyomi.databinding.RecyclerWithScrollerBinding
 import eu.kanade.tachiyomi.extension.model.Extension
 import eu.kanade.tachiyomi.extension.model.InstallStep
 import eu.kanade.tachiyomi.extension.model.InstalledExtensionsOrder
+import eu.kanade.tachiyomi.extension.util.ExtensionInstaller
+import eu.kanade.tachiyomi.extension.util.ExtensionLoader
 import eu.kanade.tachiyomi.ui.extension.details.ExtensionDetailsController
 import eu.kanade.tachiyomi.ui.main.MainActivity
 import eu.kanade.tachiyomi.ui.migration.BaseMigrationInterface
@@ -30,6 +32,7 @@ import eu.kanade.tachiyomi.ui.migration.SourceAdapter
 import eu.kanade.tachiyomi.ui.migration.SourceItem
 import eu.kanade.tachiyomi.ui.migration.manga.design.PreMigrationController
 import eu.kanade.tachiyomi.ui.source.BrowseController
+import eu.kanade.tachiyomi.util.system.isPackageInstalled
 import eu.kanade.tachiyomi.util.system.materialAlertDialog
 import eu.kanade.tachiyomi.util.system.rootWindowInsetsCompat
 import eu.kanade.tachiyomi.util.view.activityBinding
@@ -220,7 +223,7 @@ class ExtensionBottomSheet @JvmOverloads constructor(context: Context, attrs: At
 
     override fun onUpdateAllClicked(position: Int) {
         (controller.activity as? MainActivity)?.showNotificationPermissionPrompt()
-        if (!presenter.preferences.useShizukuForExtensions() &&
+        if (presenter.preferences.extensionInstaller().get() != ExtensionInstaller.SHIZUKU &&
             !presenter.preferences.hasPromptedBeforeUpdateAll().get()
         ) {
             controller.activity!!.materialAlertDialog()
@@ -410,7 +413,24 @@ class ExtensionBottomSheet @JvmOverloads constructor(context: Context, attrs: At
     }
 
     override fun uninstallExtension(pkgName: String) {
-        presenter.uninstallExtension(pkgName)
+        if (context.isPackageInstalled(pkgName)) {
+            presenter.uninstallExtension(pkgName)
+        } else {
+            val extName = run {
+                val appInfo = ExtensionLoader.getExtensionPackageInfoFromPkgName(
+                    context,
+                    pkgName,
+                )?.applicationInfo ?: return@run pkgName
+                context.packageManager.getApplicationLabel(appInfo).toString()
+            }
+            controller.activity!!.materialAlertDialog()
+                .setTitle(extName)
+                .setPositiveButton(R.string.uninstall) { _, _ ->
+                    presenter.uninstallExtension(pkgName)
+                }
+                .setNegativeButton(android.R.string.cancel) { _, _ -> }
+                .show()
+        }
     }
 
     fun onDestroy() {
